Death Condition
Added a death system. A quick overview:
Created a deathscreen.tscn to handle the visuals and clickables for the death. Creates a child of this scene under the player when he dies. This scene tweens a camera zoom and slows game time to a freeze. Adds an interactable button to restart the game. Added a death animation. More stuff I probably forgot.